+++ src/main/java/at/fhtw/swen3/persistence/entity/NewParcelInfo.java
@@ -1,4 +1,4 @@
-package at.fhtw.swen3.persistence;
+package at.fhtw.swen3.persistence.entity;
 
 import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
