+++ src/main/java/at/fhtw/swen3/persistence/entity/Parcel.java
@@ -1,6 +1,8 @@
-package at.fhtw.swen3.persistence;
+package at.fhtw.swen3.persistence.entity;
 
 import java.util.Objects;
+
+import at.fhtw.swen3.persistence.Recipient;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import org.springframework.validation.annotation.Validated;
