+++ src/main/java/at/fhtw/swen3/services/ParcelApi.java
@@ -6,9 +6,9 @@
 package at.fhtw.swen3.services;
 
 import at.fhtw.swen3.persistence.Error;
-import at.fhtw.swen3.persistence.NewParcelInfo;
-import at.fhtw.swen3.persistence.Parcel;
-import at.fhtw.swen3.persistence.TrackingInformation;
+import at.fhtw.swen3.persistence.entity.NewParcelInfo;
+import at.fhtw.swen3.persistence.entity.Parcel;
+import at.fhtw.swen3.persistence.entity.TrackingInformation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
